% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aabin2aastring.R
\name{aabin2aastring}
\alias{aabin2aastring}
\title{aabin2aastring}
\usage{
aabin2aastring(aabin)
}
\arguments{
\item{aabin}{\code{ape} \code{AAbin} [mandatory]}
}
\value{
An object of class \code{AAStringSet}
}
\description{
This function converts an \code{ape} \code{AAbin} into
\code{AAStringSet}.
}
\examples{
data(woodmouse, package="ape")
## convert into AAStringSet
#aabin2aastring(ape::trans(woodmouse, 2))
ape::trans(woodmouse, 2) |> aabin2aastring()
}
\seealso{
\code{\link[seqinr]{as.alignment}}
\code{\link[ape]{as.DNAbin.alignment}}
\code{\link[Biostrings]{AAStringSet}}
}
\author{
Kristian K Ullrich
}
