% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAMatrix-data.R
\docType{data}
\name{AAMatrix-data}
\alias{AAMatrix-data}
\alias{AAMatrix}
\title{AAMatrix-data}
\format{
an object of class \code{matrix}
}
\usage{
data(AAMatrix)
}
\value{
\code{score matrix}
}
\description{
\code{getAAMatrix()}  from the \code{alakazam} package.
}
\examples{
data("AAMatrix", package="MSA2dist")
}
\references{
Gupta N, Vander Heiden J, Uduman M, Gadala-Maria D, Yaari G,
Kleinstein S (2015) Change-O: a toolkit for analyzing large-scale B cell
immunoglobulin repertoire sequencing data. \emph{Bioinformatics.}
\bold{31(20)}, 3356-3358.
}
\keyword{datasets}
