% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon2numberAMBIG.R
\name{codon2numberAMBIG}
\alias{codon2numberAMBIG}
\title{codon2numberAMBIG}
\usage{
codon2numberAMBIG(codon)
}
\arguments{
\item{codon}{[mandatory]}
}
\value{
An object of class \code{numeric}
}
\description{
This function converts a \code{codon} into a \code{number},
but accept N and -.
}
\examples{
#unlist(lapply(names(Biostrings::GENETIC_CODE), codon2numberAMBIG))
names(Biostrings::GENETIC_CODE) |> codon2numberAMBIG()
}
\seealso{
\code{\link[Biostrings]{GENETIC_CODE}}
}
\author{
Kristian K Ullrich
}
