% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon2numberTCAG.R
\name{codon2numberTCAG}
\alias{codon2numberTCAG}
\title{codon2numberTCAG}
\usage{
codon2numberTCAG(codon)
}
\arguments{
\item{codon}{[mandatory]}
}
\value{
An object of class \code{numeric}
}
\description{
This function converts a \code{codon} into a \code{number}.
}
\examples{
#unlist(lapply(names(Biostrings::GENETIC_CODE), codon2numberTCAG))
names(Biostrings::GENETIC_CODE) |> codon2numberTCAG()
}
\seealso{
\code{\link[Biostrings]{GENETIC_CODE}}
}
\author{
Kristian K Ullrich
}
