% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseChar.R
\name{collapseChar}
\alias{collapseChar}
\title{collapseChar}
\usage{
collapseChar(xchar)
}
\arguments{
\item{xchar}{\code{character vector} [mandatory]}
}
\value{
list containing:
\itemize{
    \item unique
    \item xmap
}
}
\description{
This function collapse identical character
sequences into unique patterns with a mapping back to the
original order.
}
\examples{
data(woodmouse, package="ape")
#collapseChar(as.character(dnabin2dnastring(woodmouse)))
woodmouse |> dnabin2dnastring() |> as.character() |>
collapseChar()
}
\author{
Kristian K Ullrich
}
