% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseXStringSet.R
\name{collapseXStringSet}
\alias{collapseXStringSet}
\title{collapseXStringSet}
\usage{
collapseXStringSet(xstringset)
}
\arguments{
\item{xstringset}{\code{XStringSet} [mandatory]}
}
\value{
list containing:
\itemize{
    \item unique
    \item xmap
}
}
\description{
This function collapse identical sequences into
unique patterns with a mapping back to original order from
a \code{BStringSet}, \code{DNAStringSet}, \code{RNAStringSet}
or \code{AAStringSet}.
}
\examples{
data(woodmouse, package="ape")
#collapseXStringSet(dnabin2dnastring(woodmouse))
woodmouse |> dnabin2dnastring() |> collapseXStringSet()
}
\seealso{
\code{\link[Biostrings]{XStringSet-class}}
}
\author{
Kristian K Ullrich
}
