% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2dxy.R
\name{dist2dxy}
\alias{dist2dxy}
\title{dist2dxy}
\usage{
dist2dxy(d, pop = NULL, popx = NULL, popy = NULL, popout = NULL)
}
\value{
list containing:
\itemize{
    \item dxy pairwise matrix
    \item fst pairwise matrix
    \item pi_within pop matrix
    \item pi_between 3d matrix
    \item valid_sites 
    \item populations
    \item popx
    \item popy
    \item popout
    \item Dx
    \item Dy
    \item Dxy
    \item FSTxy
    \item Dmin
    \item Dmax
    \item Dx_out
    \item Dy_out
    \item Dout
    \item Gmin
    \item RND
    \item RNDmin
    \item RNDmax
    \item regionUsed
    \item pop.info
}
}
\description{
This function takes the results of the
\code{dnastring2dist} function to calculate dxy and fst for all
population combinations and optional RND, Gmin, RNDmin and RNDmax
if one specifies \code{popx}, \code{popy} and \code{popout}.
}
\details{
The function computes pairwise genetic differentiation statistics
from a precomputed sequence distance matrix.

Let $d_{ij}$ denote the per-site sequence distance between sequences
$i$ and $j$, and let $S_{ij}$ be the number of aligned sites used.

\bold{Within-population diversity}

For population $x$ with sequences $i in x$:

\deqn{
D_x = \frac{1}{\binom{n_X}{2}} \sum_{i<j in X} d_{ij}
}{D_x = mean pairwise distance within population x}

\bold{Between-population divergence (Dxy)}

For populations $x$ and $y$:

\deqn{
D_{xy} = \frac{1}{n_x n_y} \sum_{i in x} \sum_{j in y} d_{ij}
}{Dxy = mean pairwise distance between populations x and y}

\bold{Fixation index (Fst)}

Using the distance-based estimator:

\deqn{
F_{ST} = \frac{D_{xy} - \frac{D_x + D_y}{2}}{D_{xy}}
}{FST = (Dxy - (Dx + Dy)/2) / Dxy}

\bold{Minimum divergence (Dmin)}

\deqn{
D_{min}(x,y) = \min_{i in x, j in y} d_{ij}
}{Dmin = minimum pairwise distance between populations}

\bold{Outgroup-corrected divergence}

Let $o$ be an outgroup population:

\deqn{
D(x,o) = \frac{1}{n_x n_o} \sum_{i in x} \sum_{k in o} d_{ik}
}
\deqn{
D(y,o) = \frac{1}{n_y n_o} \sum_{j in y} \sum_{k in o} d_{jk}
}
\deqn{
D_{out} = \frac{D(x,o) + D(y,o)}{2}
}

\bold{RND statistics}

\deqn{
RND = \frac{D_{xy}}{D_{out}}
}{RND = Dxy / Dout}

\deqn{
G_{min} = \frac{D_{min}(x,y)}{D_{xy}}
}{Gmin = Dmin / Dxy}

\deqn{
RND_{min} = \frac{D_{min}(x,y)}{D_{out}}
}{RNDmin = Dmin / Dout}

\bold{Interpretation}
\itemize{
    \item $D_{xy}$: average sequence divergence between populations
    \item $F_{ST}$: relative differentiation scaled by
within-population diversity
    \item $D_{min}$: closest haplotype distance between populations
    \item $RND$: divergence normalized by outgroup distance
    \item $G_{min}$: proportion of minimum divergence relative to
mean divergence
    \item $RND_{min}$: minimum divergence scaled by outgroup distance
    \item $RND_{max}$: max divergence scaled by outgroup distance
}
}
\examples{
## load example sequence data
data("iupac", package="MSA2dist")
#d <- dnastring2dist(iupac)
#dist2dxy(d)
iupac |> dnastring2dist() |> dist2dxy()
## create poplist
poplist <- list(FRA = grep("Mmd.FRA", names(iupac)),
    GER = grep("Mmd.GER", names(iupac)),
    IRA = grep("Mmd.IRA", names(iupac)),
    AFG = grep("Mmm.AFG", names(iupac)))
iupac <- iupac |> addpop2string(poplist)
#d <- dnastring2dist(iupac)
#dist2dxy(d, pop=popinteger(iupac))
iupac |> dnastring2dist() |> dist2dxy(pop=popinteger(iupac))
#d <- dnastring2dist(iupac)
#dist2dxy(d, pop=popinteger(iupac),
#popx="FRA", popy="IRA", popout="AFG")
iupac |> dnastring2dist() |> dist2dxy(pop=popinteger(iupac),
popx="FRA", popy="IRA", popout="AFG")
}
\seealso{
\code{\link[MSA2dist]{dnastring2dist}},
\code{\link[MSA2dist]{addpop2string}}
}
\author{
Kristian K Ullrich
}
