% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin2dnastring.R
\name{dnabin2dnastring}
\alias{dnabin2dnastring}
\title{dnabin2dnastring}
\usage{
dnabin2dnastring(dnabin)
}
\arguments{
\item{dnabin}{\code{ape} \code{DNAbin} [mandatory]}
}
\value{
An object of class \code{DNAStringSet}
}
\description{
This function converts an \code{ape} \code{DNAbin} into a
\code{DNAStringSet}.
}
\examples{
data(woodmouse, package="ape")
## convert into DNAStringSet
#dnabin2dnastring(woodmouse)
woodmouse |> dnabin2dnastring()
}
\seealso{
\code{\link[seqinr]{as.alignment}}
\code{\link[ape]{as.DNAbin.alignment}}
\code{\link[Biostrings]{DNAStringSet}}
}
\author{
Kristian K Ullrich
}
