% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnastring2kaks.R
\name{dnastring2kaks}
\alias{dnastring2kaks}
\title{dnastring2kaks}
\usage{
dnastring2kaks(cds, model = "Li", threads = 1)
}
\arguments{
\item{cds}{\code{DNAStringSet} coding sequence alignment [mandatory]}

\item{model}{specify codon model either "Li" or "NG86" [default: Li]}

\item{threads}{number of parallel threads [default: 1]}
}
\value{
A \code{data.frame} of \code{KaKs} values
}
\description{
This function calculates Ka/Ks (pN/pS; according to
\emph{Li (1993)} or \emph{Nei and Gojobori (1986)} for all combinations of
a \code{DNAStringSet}.
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
#dnastring2kaks(hiv, model="Li")
hiv |> dnastring2kaks(model="Li")
#dnastring2kaks(hiv, model="NG86")
hiv |> dnastring2kaks(model="NG86")
#dnastring2kaks(hiv, model="NG86", threads=2)
hiv |> dnastring2kaks(model="NG86", threads=2)
}
\references{
Nei and Gojobori. (1986) Simple methods for estimating the
numbers of synonymous and nonsynonymous nucleotide substitutions.
\emph{Mol. Biol. Evol.}, \bold{3(5)}, 418-426.

Ganeshan et al. (1997) Human immunodeficiency virus type 1
genetic evolution in children with different rates of development of disease.
\emph{J. Virology.} \bold{71(1)}, 663-677.

Yang et al. (2000) Codon-substitution models for heterogeneous
selection pressure at amino acid sites. \emph{Genetics.}
\bold{155(1)}, 431-449.
}
\seealso{
\code{\link[seqinr]{kaks}}
}
\author{
Kristian K Ullrich
}
