% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnastring2weightedPi.R
\name{dnastring2weightedPi}
\alias{dnastring2weightedPi}
\title{dnastring2weightedPi}
\usage{
dnastring2weightedPi(
    dna,
    model = "IUPAC",
    estimator = "count",
    mask = NULL,
    region = NULL,
    pop = NULL
)
}
\arguments{
\item{dna}{\code{DNAStringSet} [mandatory]}

\item{model}{specify model either "IUPAC" (diploid; decode IUPAC) or
"sequence" (haploid; only A,C,T,G counted as discrete alleles)
[default: IUPAC]}

\item{estimator}{specify estimator either
"count" (count-based; computes nucleotide diversity as the observed
proportion of pairwise allele differences, nDiff / nComp) or
"prob" (allele-frequency based; computes expected heterozygosity
as 1 - sum p_i^2) [default: count]}

\item{mask}{\code{IRanges} object indicating masked sites
[default: NULL]}

\item{region}{\code{IRanges} object indicating region to use for dist
calculation. Default is null, meaning all sites are used [default: NULL]}

\item{pop}{[default: NULL] optional population assignment:
\itemize{
    \item character vector (same length as dna)
    \item named character vector (mapping seqnames to pop)
    \item integer vector (same length as dna)
    \item named integer vector (mapping index to pop)
    \item list (see \code{addpop2dnastring()})
}}
}
\value{
list containing:
\itemize{
    \item weightedPi
    \item populations
    \item model
    \item estimator
    \item regionUsed
    \item pop.info
}
}
\description{
This function calculates weightedPi for all
population combinations of a \code{DNAStringSet}.
}
\examples{
## load example sequence data
data("iupac", package="MSA2dist")
#dnastring2weightedPi(iupac, model="IUPAC")
iupac |> dnastring2weightedPi(model="IUPAC")
#dnastring2weightedPi(iupac, model="IUPAC", estimator="prob")
iupac |> dnastring2weightedPi(model="IUPAC", estimator="prob")
#dnastring2weightedPi(iupac, model="sequence")
iupac |> dnastring2weightedPi(model="sequence")
#dnastring2weightedPi(iupac, model="sequence", estimator="prob")
iupac |> dnastring2weightedPi(model="sequence", estimator="prob")
## create poplist
poplist <- list(FRA = grep("Mmd.FRA", names(iupac)),
    GER = grep("Mmd.GER", names(iupac)),
    IRA = grep("Mmd.IRA", names(iupac)),
    AFG = grep("Mmm.AFG", names(iupac)))
iupac <- iupac |> addpop2string(poplist)
#dnastring2weightedPi(iupac, pop=popinteger(iupac), model="IUPAC")
iupac |> dnastring2weightedPi(pop=popinteger(iupac), model="IUPAC")
#dnastring2weightedPi(iupac, pop=popnames(iupac), model="IUPAC")
iupac |> dnastring2weightedPi(pop=popnames(iupac), model="IUPAC")
#dnastring2weightedPi(iupac, pop=popinteger(iupac), model="IUPAC",
#estimator="prob")
iupac |> dnastring2weightedPi(pop=popinteger(iupac), model="IUPAC",
estimator="prob")
#dnastring2weightedPi(iupac, pop=popinteger(iupac), model="sequence")
iupac |> dnastring2weightedPi(pop=popinteger(iupac), model="sequence")
#dnastring2weightedPi(iupac, pop=popinteger(iupac), model="sequence",
#estimator="prob")
iupac |> dnastring2weightedPi(pop=popinteger(iupac), model="sequence",
estimator="prob")
## create mask
mask1 <- IRanges::IRanges(start=c(1,61,121), end=c(30,90,150))
## use mask
iupac |> dnastring2weightedPi(model="IUPAC", mask=mask1)
## use region
region1 <- IRanges::IRanges(start=c(1,139), end=c(75,225))
iupac |> dnastring2weightedPi(model="IUPAC", region=region1)
## use mask and region
iupac |> dnastring2weightedPi(model="IUPAC", mask=mask1, region=region1)
}
\seealso{
\code{\link[MSA2dist]{dnastring2dist}},
\code{\link[MSA2dist]{addpop2string}}
}
\author{
Kristian K Ullrich
}
