% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENETIC_CODE_TCAG.R
\name{GENETIC_CODE_TCAG}
\alias{GENETIC_CODE_TCAG}
\alias{codon2number}
\title{GENETIC_CODE_TCAG}
\usage{
codon2number(codon)
}
\arguments{
\item{codon}{\code{codon} [mandatory]}
}
\value{
An object of class \code{numeric}
}
\description{
\code{GENETIC_CODE} from \code{Biostrings} extended by codon
number and number of syn sites.
}
\examples{
GENETIC_CODE_TCAG
}
\seealso{
\code{\link[Biostrings]{GENETIC_CODE}}
}
\author{
Kristian K Ullrich
}
