% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv-data.R
\docType{data}
\name{hiv-data}
\alias{hiv-data}
\alias{hiv}
\title{hiv-data}
\format{
an object of class \code{DNAStringSet}
see \code{\link[Biostrings]{XStringSet-class}}
}
\usage{
data(hiv)
}
\description{
Example cds sequences from HIV-1 sample 136 patient 1 from
Sweden envelope glycoprotein (env) gene, V3 region as \code{DNAStringSet}.
}
\examples{
data("hiv", package="MSA2dist")
}
\references{
Yang et al. (2000) Codon-substitution models for heterogeneous
selection pressure at amino acid sites. \emph{Genetics.}
\bold{155(1)}, 431-449.
}
\keyword{datasets}
