% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupac-data.R
\docType{data}
\name{iupac-data}
\alias{iupac-data}
\alias{iupac}
\title{iupac-data}
\format{
an object of class \code{DNAStringSet}
see \code{\link[Biostrings]{XStringSet-class}}
}
\usage{
data(iupac)
}
\description{
Example IUPAC sequences created with \code{angsd}
from different house mouse (Mus musculus) sub-populations
from Harr et al. (2016) \code{DNAStringSet}.
}
\examples{
data("iupac", package="MSA2dist")
}
\references{
Harr et al. (2016) Genomic resources for wild populations of the
house mouse, Mus musculus and its close relative
Mus spretus. \emph{Scientific data.} \bold{3(1)}, 1-14.
}
\keyword{datasets}
