% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupacMatrix.R
\name{iupacMatrix}
\alias{iupacMatrix}
\title{iupacMatrix}
\usage{
iupacMatrix()
}
\value{
\code{score matrix}
}
\description{
This function creates a \code{iupacMatrix} object to be used
with the \code{rcpp_distSTRING} function. By default, the \code{iupac matrix}
is defined as literal distance obtained from \code{Chang et al. 2017}.
(see \url{https://link.springer.com/article/10.1007/s00335-017-9704-9})
}
\examples{
iupacMatrix()
}
\references{
Chang,P. L.,Kopania,E.,Keeble,S.,Sarver,B. A.,Larson,
E.,Orth,A.,... & Dean,M. D. (2017). Whole exome sequencing of
wild-derived inbred strains of mice improves power to link phenotype and
genotype. \emph{Mammalian genome},\bold{28(9-10)},416-425.
}
\seealso{
\link[MSA2dist]{dnastring2dist},\link[ape]{dist.dna}
}
\author{
Kristian K Ullrich
}
