% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePostalignedSeqs.R
\name{makePostalignedSeqs}
\alias{makePostalignedSeqs}
\title{makePostalignedSeqs}
\usage{
makePostalignedSeqs(x)
}
\arguments{
\item{x}{x}
}
\value{
get internal function makePostalignedSeqs
}
\description{
This function is a fork from an internal function from
\code{Biostrings}
}
\examples{
## define two cds sequences
cds1 <- Biostrings::DNAString("ATGCAACATTGC")
cds2 <- Biostrings::DNAString("ATGCATTGC")
makePostalignedSeqs(Biostrings::pairwiseAlignment(
    cds2aa(Biostrings::DNAStringSet(cds1)),
    cds2aa(Biostrings::DNAStringSet(cds2))))
}
\seealso{
\code{\link[Biostrings]{pairwiseAlignment}},
\link[MSA2dist]{cds2codonaln}
}
\author{
Kristian K Ullrich
}
