% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normpop.R
\name{normpop}
\alias{normpop}
\title{normpop}
\usage{
normpop(pop, seqnames)
}
\arguments{
\item{pop}{[mandatory] population assignment:
\itemize{
    \item character vector (same length as seqnames)
    \item named character vector (mapping seqnames to pop)
    \item integer vector (same length as seqnames)
    \item named integer vector (mapping index to pop)
    \item list (see \code{addpop2dnastring()})
}}

\item{seqnames}{sequence names [mandatory]}
}
\value{
list containing:
\itemize{
    \item pop_idx
    \item pop_name
    \item pop_levels
    \item pop_lookup
    \item poplist.names
    \item poplist.integer
}
}
\description{
This function normalizes population representation
}
\examples{
## load example sequence data
data("iupac", package="MSA2dist")
## create poplist
poplist <- list(FRA = grep("Mmd.FRA", names(iupac)),
    GER = grep("Mmd.GER", names(iupac)),
    IRA = grep("Mmd.IRA", names(iupac)),
    AFG = grep("Mmm.AFG", names(iupac)))
normpop(pop=poplist, seqnames=names(iupac))
## mxixing index and names
poplist <- list(FRA = names(iupac)[grep("Mmd.FRA", names(iupac))],
    GER = grep("Mmd.GER", names(iupac)),
    IRA = names(iupac)[grep("Mmd.IRA", names(iupac))],
    AFG = grep("Mmm.AFG", names(iupac)))
normpop(pop=poplist, seqnames=names(iupac))
## leaving out some sequences which will be assigned as "unassigned"
poplist <- list(FRA = names(iupac)[grep("Mmd.FRA", names(iupac))],
    GER = grep("Mmd.GER", names(iupac)),
    IRA = names(iupac)[grep("Mmd.IRA", names(iupac))])
normpop(pop=poplist, seqnames=names(iupac))
## create character vector
popchar <- c(rep("FRA", 8),
    rep("GER", 8),
    rep("IRA", 8),
    rep("AFG", 6))
normpop(pop=popchar, seqnames=names(iupac))
## create named character vector
popcharn <- c(rep("FRA", 8),
    rep("GER", 8),
    rep("IRA", 8),
    rep("AFG", 6))
names(popcharn) <- names(iupac)
normpop(pop=popcharn, seqnames=names(iupac))
## create integer vector
popint <- c(rep(1L, 8),
    rep(2L, 8),
    rep(3L, 8),
    rep(4L, 6))
normpop(pop=popint, seqnames=names(iupac))
## create named integer vector
popintn <- c(rep(1L, 8),
    rep(2L, 8),
    rep(3L, 8),
    rep(4L, 6))
names(popintn) <- names(iupac)
normpop(pop=popintn, seqnames=names(iupac))
}
\seealso{
\code{\link[MSA2dist]{addpop2string}}
}
\author{
Kristian K Ullrich
}
