% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_distSTRING}
\alias{rcpp_distSTRING}
\title{rcpp_distSTRING}
\usage{
rcpp_distSTRING(dnavector, scoreMatrix, ncores = 1L, symmetric = 1L)
}
\arguments{
\item{dnavector}{StringVector [mandatory]}

\item{scoreMatrix}{NumericMatrix [mandatory]}

\item{ncores}{number of cores [default: 1]}

\item{symmetric}{symmetric score matrix [default: 1]}
}
\value{
list containing:
\itemize{
    \item distSTRING
    \item sitesUsed
}
}
\description{
calculates pairwise distances using a score matrix
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
rcpp_distSTRING(dnavector=as.character(hiv), scoreMatrix=iupacMatrix())
}
\author{
Kristian K Ullrich
}
