% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_dxy_fst_pop}
\alias{rcpp_dxy_fst_pop}
\title{rcpp_dxy_fst_pop}
\usage{
rcpp_dxy_fst_pop(dnavector, pop_idx, model = "IUPAC", estimator = "prob")
}
\arguments{
\item{dnavector}{StringVector [mandatory]}

\item{pop_idx}{IntegerVector [mandatory]}

\item{model}{string [default: IUPAC]}

\item{estimator}{string [default: count]}
}
\value{
list containing:
\itemize{
    \item dxy
    \item fst
    \item pi_within
    \item pi_between
    \item populations
    \item model
    \item estimator
}
}
\description{
calculates dxy and fst per population combination
}
\examples{
## load example sequence data
data("iupac", package="MSA2dist")
rcpp_dxy_fst_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="IUPAC",
estimator="count")
rcpp_dxy_fst_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="IUPAC",
estimator="prob")
rcpp_dxy_fst_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="sequence",
estimator="count")
rcpp_dxy_fst_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="sequence",
estimator="prob")
}
\author{
Kristian K Ullrich
}
