% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_pairwiseDeletionDNA}
\alias{rcpp_pairwiseDeletionDNA}
\title{rcpp_pairwiseDeletionDNA}
\usage{
rcpp_pairwiseDeletionDNA(dnavector, ncores = 1L, symmetric = 1L)
}
\arguments{
\item{dnavector}{StringVector [mandatory]}

\item{ncores}{number of cores [default: 1]}

\item{symmetric}{symmetric score matrix [default: 1]}
}
\value{
list containing:
\itemize{
    \item sitesUsed
}
}
\description{
returns number of DNA sites used
}
\examples{
## load example sequence data
data("woodmouse", package="ape")
w <- woodmouse |> dnabin2dnastring() |> as.character()
rcpp_pairwiseDeletionDNA(dnavector=w, ncores=1)
}
\author{
Kristian K Ullrich
}
