% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_weightedPi}
\alias{rcpp_weightedPi}
\title{rcpp_weightedPi}
\usage{
rcpp_weightedPi(dnavector, model = "IUPAC", estimator = "count")
}
\arguments{
\item{dnavector}{StringVector [mandatory]}

\item{model}{string [default: IUPAC]}

\item{estimator}{string [default: count]}
}
\value{
list containing:
\itemize{
    \item weightedPi
    \item model
    \item estimator
}
}
\description{
calculates weighted pi
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
rcpp_weightedPi(dnavector=as.character(hiv), model="IUPAC",
estimator="count")
rcpp_weightedPi(dnavector=as.character(hiv), model="sequence",
estimator="count")
rcpp_weightedPi(dnavector=as.character(hiv), model="IUPAC",
estimator="prob")
rcpp_weightedPi(dnavector=as.character(hiv), model="sequence",
estimator="prob")
}
\author{
Kristian K Ullrich
}
