% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_weightedPi_pop}
\alias{rcpp_weightedPi_pop}
\title{rcpp_weightedPi_pop}
\usage{
rcpp_weightedPi_pop(dnavector, pop_idx, model = "IUPAC", estimator = "count")
}
\arguments{
\item{dnavector}{StringVector [mandatory]}

\item{pop_idx}{IntegerVector [mandatory]}

\item{model}{string [default: IUPAC]}

\item{estimator}{string [default: count]}
}
\value{
list containing:
\itemize{
    \item weightedPi
    \item populations
    \item model
    \item estimator
}
}
\description{
calculates weighted pi per population
}
\examples{
## load example sequence data
data("iupac", package="MSA2dist")
rcpp_weightedPi_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="IUPAC",
estimator="count")
rcpp_weightedPi_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="IUPAC",
estimator="prob")
rcpp_weightedPi_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="sequence",
estimator="count")
rcpp_weightedPi_pop(dnavector=as.character(iupac),
pop_idx=c(rep(1,8),rep(2,8),rep(3,8),rep(4,6)),
model="sequence",
estimator="prob")
}
\author{
Kristian K Ullrich
}
