% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptriidx.R
\name{uptriidx}
\alias{uptriidx}
\title{uptriidx}
\usage{
uptriidx(n, diag = FALSE)
}
\arguments{
\item{n}{dimension of initial matrix [mandatory]}

\item{diag}{indicate if diag should be retained [default: FALSE]}
}
\value{
list of positions
}
\description{
This function returns upper tri index for usage with
\code{pivot_long} reduction.
}
\examples{
uptriidx(10)
}
\author{
Kristian K Ullrich
}
