% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_pairwiseDeletionAA}
\alias{rcpp_pairwiseDeletionAA}
\title{rcpp_pairwiseDeletionAA}
\usage{
rcpp_pairwiseDeletionAA(aavector, ncores = 1L, symmetric = 1L)
}
\arguments{
\item{aavector}{StringVector [mandatory]}

\item{ncores}{number of cores [default: 1]}

\item{symmetric}{symmetric score matrix [default: 1]}
}
\value{
list
}
\description{
returns number of AA sites used
}
\examples{
## load example sequence data
data("hiv", package="MSA2dist")
h <- hiv |> cds2aa() |> as.character()
rcpp_pairwiseDeletionAA(aavector=h, ncores=1)
}
\author{
Kristian K Ullrich
}
